refactor: account_delegation_list query
