+++ x/registry/keeper/grpc_query_account_delegation_list.go
@@ -1,50 +1,47 @@
 package keeper
 
 import (
+	"bytes"
 	"context"
-	"github.com/cosmos/cosmos-sdk/store/prefix"
-	sdkErrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/query"
 
 	"github.com/KYVENetwork/chain/x/registry/types"
+	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/query"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
-// DelegatorsByPoolAndStaker returns all delegators for a specific pool that delegated to the given staker address
-// It also returns the pool-object and the delegation information for that pool
+// AccountDelegationList returns all staker with their pools the given user has delegated to.
+// It calculates the current rewards.
 // Pagination is supported
-func (k Keeper) DelegatorsByPoolAndStaker(goCtx context.Context, req *types.QueryDelegatorsByPoolAndStakerRequest) (*types.QueryDelegatorsByPoolAndStakerResponse, error) {
+func (k Keeper) AccountDelegationList(goCtx context.Context, req *types.QueryAccountDelegationListRequest) (*types.QueryAccountDelegationListResponse, error) {
 	if req == nil {
 		return nil, status.Error(codes.InvalidArgument, "invalid request")
 	}
 
+	var delegated []types.DelegatorResponse
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
-	pool, found := k.GetPool(ctx, req.PoolId)
-	if !found {
-		return nil, sdkErrors.Wrapf(sdkErrors.ErrNotFound, types.ErrPoolNotFound.Error(), req.PoolId)
-	}
-
-	var delegators []types.StakerDelegatorResponse
-
 	store := ctx.KVStore(k.storeKey)
-	// Build prefix. Store is already indexed in an optimal way
-	prefixBuilder := types.KeyPrefixBuilder{Key: types.KeyPrefix(types.DelegatorKeyPrefix)}.AInt(pool.Id).AString(req.Staker).Key
-	delegatorStore := prefix.NewStore(store, prefixBuilder)
+	delegatorStore := prefix.NewStore(store, types.KeyPrefix(types.DelegatorKeyPrefix))
 
+	// TODO find indexing solution for performance
 	pageRes, err := query.FilteredPaginate(delegatorStore, req.Pagination, func(key []byte, value []byte, accumulate bool) (bool, error) {
 
 		if accumulate {
+			// Check if entry belongs to given address (delegator)
+			if bytes.Compare(key[53:96], []byte(req.Address)) != 0 {
+				return false, nil
+			}
 
 			var delegator types.Delegator
-
 			if err := k.cdc.Unmarshal(value, &delegator); err != nil {
 				return false, nil
 			}
 
-			// Calculate current rewards for the delegator
+			pool, _ := k.GetPool(ctx, delegator.Id)
+
 			f1 := F1Distribution{
 				k:                k,
 				ctx:              ctx,
@@ -53,11 +50,15 @@ func (k Keeper) DelegatorsByPoolAndStaker(goCtx context.Context, req *types.Quer
 				delegatorAddress: delegator.Delegator,
 			}
 
-			delegators = append(delegators, types.StakerDelegatorResponse{
-				Delegator:        delegator.Delegator,
-				CurrentReward:    f1.getCurrentReward(),
-				DelegationAmount: delegator.DelegationAmount,
-				Staker:           req.Staker,
+			delegationPoolData, _ := k.GetDelegationPoolData(ctx, pool.Id, delegator.Staker)
+
+			delegated = append(delegated, types.DelegatorResponse{
+				Account:            req.Address,
+				Pool:               &pool,
+				CurrentReward:      f1.getCurrentReward(),
+				DelegationAmount:   delegator.DelegationAmount,
+				Staker:             delegator.Staker,
+				DelegationPoolData: &delegationPoolData,
 			})
 		}
 
@@ -68,12 +69,8 @@ func (k Keeper) DelegatorsByPoolAndStaker(goCtx context.Context, req *types.Quer
 		return nil, status.Error(codes.Internal, err.Error())
 	}
 
-	delegationPoolData, _ := k.GetDelegationPoolData(ctx, pool.Id, req.Staker)
-
-	return &types.QueryDelegatorsByPoolAndStakerResponse{
-		Delegators:         delegators,
-		Pool:               &pool,
-		DelegationPoolData: &delegationPoolData,
-		Pagination:         pageRes,
+	return &types.QueryAccountDelegationListResponse{
+		Delegations: delegated,
+		Pagination:  pageRes,
 	}, nil
 }
